backend/requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.11.0
djangorestframework==3.12.4
gunicorn==20.1.0
hideme==0.0.2
idna==2.10
psycopg2==2.8.6
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
